Moved back hdx api
